Update main.lua
This gives the ability to return amount of items available as well as the slot. You Can also request items between slots. You can also just check a specific amount of slots for example just check the first 5 slots or can check from slot 6 to 25 etc. This also gives you the ability to send the slot number to the server and remove that item if the itemname and slot exists which will help with performance to not need to loop the pockets on server for each sale. This is something I am using in the hunting script to make it more efficient when selling items to butcher.

Examples:
local amount, slot = exports['qbr-inventory']:GetItemAmount(itemname)

local data = exports['qbr-inventory']:GetSlotData(1, 5) -- This will return the data for slots 1 - 5